Avoid unnecessary precision

+++ galaxies.py
@@ -59,6 +59,8 @@ def combine(gal_seds_in, gal_seds_out, el, pa, re, sersic):
     gal_seds_out = gal_seds_out[:,None,None]
     gal_input = gal_images * gal_seds_in
     gal_target = gal_images * gal_seds_out
+    gal_input = gal_input.astype(np.float32)
+    gal_target = gal_target.astype(np.float32)
     return gal_input, gal_target
 
 
@@ -93,8 +95,8 @@ def main(n=10000):
     gal_target = np.delete(gal_target, np.arange(1,17,2), axis = 3)
 
     # making input and target redshift arrays
-    z_in = z[z_in_idx]
-    z_out = z[z_out_idx]
+    z_in = z[z_in_idx].astype(np.float32)
+    z_out = z[z_out_idx].astype(np.float32)
     
     ## saving input and target galaxies to npy files
     np.save('inputgalaxies.npy', gal_input)
